parser/tests/tree_construction/basic.dat

#data
<p><font size=4><font color=red><font size=4><font size=4><font size=4><font size=4><font size=4><font color=red><p>X
#errors
(1,3): expected-doctype-but-got-start-tag
(1,116): unexpected-end-tag
(1,117): expected-closing-tag-but-got-eof
#document
| <html>
|   <head>
|   <body>
|     <p>
|       <font>
|         size="4"
|         <font>
|           color="red"
|           <font>
|             size="4"
|             <font>
|               size="4"
|               <font>
|                 size="4"
|                 <font>
|                   size="4"
|                   <font>
|                     size="4"
|                     <font>
|                       color="red"
|     <p>
|       <font>
|         color="red"
|         <font>
|           size="4"
|           <font>
|             size="4"
|             <font>
|               size="4"
|               <font>
|                 color="red"
|                 "X"
